pyproject.toml
[project]
name = "journals-keyword-searching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "aiohttp",
    "tqdm",
    "stanza>=1.11.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pdfminer-six>=20251230",
]
